backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.1